Håndterer at bransjer kan mangle for arbeidsgiver
- Takler henting av data for arbeidsgiver med null referanse.
- Upserter bransjer for arbeidsgivere som mangler bransje ved oppdatering.

+++ src/main/kotlin/no/nav/helse/modell/VedtakDao.kt
@@ -251,7 +251,7 @@ internal class VedtakDao(private val dataSource: DataSource) {
             fødselsdato = row.localDateOrNull("fodselsdato"),
             kjønn = row.stringOrNull("kjonn")?.let(Kjønn::valueOf)
         ),
-        arbeidsgiverRef = row.int("arbeidsgiver_ref"),
+        arbeidsgiverRef = row.long("arbeidsgiver_ref"),
         speilSnapshotRef = row.int("speil_snapshot_ref"),
         infotrygdutbetalingerRef = row.intOrNull("infotrygdutbetalinger_ref")
     )
