+++ src/main/kotlin/no/nav/helse/modell/vedtaksperiode/VedtaksperiodeDto.kt
@@ -6,7 +6,7 @@ data class VedtaksperiodeDto(
     val fødselsnummer: String,
     val aktørId: String,
     val personinfo: PersoninfoDto,
-    val arbeidsgiverRef: Int,
+    val arbeidsgiverRef: Long,
     val speilSnapshotRef: Int,
     val infotrygdutbetalingerRef: Int?
 )
